Interfaces, new reading method in CSV_READER.java for "RatingFiltrado".csv

+++ src/main/java/org/example/Class/Ratings.java
@@ -5,11 +5,11 @@ import lombok.Getter;
 @Getter
 public class Ratings {
     private int movieID;
-    private int rating;
+    private float rating;
     private int timestamp;
     private int UserID;
 
-    public Ratings(int movieID, int rating, int timestamp, int userID) {
+    public Ratings(int movieID, float rating, int timestamp, int userID) {
         this.movieID = movieID;
         this.rating = rating;
         this.timestamp = timestamp;
